Homework #15 finally done!
Finally done! Interpolated wind speed values along the turbine blades are now correct in meters per second. I added a 1-dimensional array to hold the wind speed values converted from knots -> meters/sec from the soundings. he conversion was done with a constant multiplied to the vector. I used this vector in the M(zr) function where the interpolation is performed. My confusion was partly with the units as I was seeing interpolations as knots and not realizing the unit error. I also did not correctly allow for the conversion from AGL to MSL heights and had to try a few attempts and logical methods to fix that problem.